+++ src/main/kotlin/teoriadosgrafos/metodosdebusca/ponderado/BellmanFord.kt
@@ -21,7 +21,7 @@ object BellmanFord {
 		val origem = origem - 1
 		/* Processo de inicialização do grafo */
 		grafo.inicializar()
-		val dist = IntArray(grafo.matrizDeAdjacencia.size) { grafo.matrizDeAdjacencia[origem][it] }
+		val dist = DoubleArray(grafo.matrizDeAdjacencia.size) { grafo.matrizDeAdjacencia[origem][it] }
 		val prev = IntArray(grafo.matrizDeAdjacencia.size) { 0 }
 		grafo.visitados[origem] = true
 		/* Conjunto de vértices do grafo */
